Extract the code for dev mode
