+++ pkg/skaffold/runner/runner_test.go
@@ -22,6 +22,9 @@ import (
 	"io/ioutil"
 	"testing"
 
+	"github.com/pkg/errors"
+	"k8s.io/client-go/tools/clientcmd/api"
+
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/local"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/tag"
@@ -30,10 +33,7 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/defaults"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/test"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/watch"
 	"github.com/GoogleContainerTools/skaffold/testutil"
-	"github.com/pkg/errors"
-	"k8s.io/client-go/tools/clientcmd/api"
 )
 
 type TestBuilder struct {
@@ -107,44 +107,6 @@ func (t *TestDeployer) Deploy(ctx context.Context, out io.Writer, builds []build
 
 func (t *TestDeployer) Cleanup(ctx context.Context, out io.Writer) error {
 	return nil
-}
-
-type TestWatcher struct {
-	changedArtifacts [][]int
-	changeCallbacks  []func(watch.Events)
-	events           []watch.Events
-	err              error
-}
-
-func NewWatcherFactory(err error, events []watch.Events, changedArtifacts ...[]int) watch.Factory {
-	return func() watch.Watcher {
-		return &TestWatcher{
-			changedArtifacts: changedArtifacts,
-			events:           events,
-			err:              err,
-		}
-	}
-}
-
-func (t *TestWatcher) Register(deps func() ([]string, error), onChange func(watch.Events)) error {
-	t.changeCallbacks = append(t.changeCallbacks, onChange)
-	return nil
-}
-
-func (t *TestWatcher) Run(ctx context.Context, trigger watch.Trigger, onChange func() error) error {
-	evts := watch.Events{}
-	if t.events != nil {
-		evts = t.events[0]
-		t.events = t.events[1:]
-	}
-
-	for _, artifactIndices := range t.changedArtifacts {
-		for _, artifactIndex := range artifactIndices {
-			t.changeCallbacks[artifactIndex](evts)
-		}
-		onChange()
-	}
-	return t.err
 }
 
 func createDefaultRunner(t *testing.T) *SkaffoldRunner {
@@ -320,181 +282,6 @@ func TestRun(t *testing.T) {
 			}})
 
 			testutil.CheckError(t, test.shouldErr, err)
-		})
-	}
-}
-
-func TestDev(t *testing.T) {
-	var tests = []struct {
-		description    string
-		builder        build.Builder
-		tester         test.Tester
-		deployer       deploy.Deployer
-		watcherFactory watch.Factory
-		shouldErr      bool
-	}{
-		{
-			description: "fails to build the first time",
-			builder: &TestBuilder{
-				errors: []error{errors.New("")},
-			},
-			tester:         &TestTester{},
-			deployer:       &TestDeployer{},
-			watcherFactory: NewWatcherFactory(nil, nil),
-			shouldErr:      true,
-		},
-		{
-			description: "fails to deploy the first time",
-			builder:     &TestBuilder{},
-			tester:      &TestTester{},
-			deployer: &TestDeployer{
-				errors: []error{errors.New("")},
-			},
-			watcherFactory: NewWatcherFactory(nil, nil),
-			shouldErr:      true,
-		},
-		{
-			description: "fails to deploy due to failed tests",
-			builder:     &TestBuilder{},
-			tester: &TestTester{
-				errors: []error{errors.New("")},
-			},
-			deployer:       &TestDeployer{},
-			watcherFactory: NewWatcherFactory(nil, nil),
-			shouldErr:      true,
-		},
-		{
-			description: "ignore subsequent build errors",
-			builder: &TestBuilder{
-				errors: []error{nil, errors.New("")},
-			},
-			tester:         &TestTester{},
-			deployer:       &TestDeployer{},
-			watcherFactory: NewWatcherFactory(nil, nil, nil),
-		},
-		{
-			description: "ignore subsequent deploy errors",
-			builder:     &TestBuilder{},
-			tester:      &TestTester{},
-			deployer: &TestDeployer{
-				errors: []error{nil, errors.New("")},
-			},
-			watcherFactory: NewWatcherFactory(nil, nil, nil),
-		},
-		{
-			description:    "fail to watch files",
-			builder:        &TestBuilder{},
-			tester:         &TestTester{},
-			deployer:       &TestDeployer{},
-			watcherFactory: NewWatcherFactory(errors.New(""), nil),
-			shouldErr:      true,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.description, func(t *testing.T) {
-			runner := createDefaultRunner(t)
-			runner.Builder = test.builder
-			runner.Tester = test.tester
-			runner.Deployer = test.deployer
-			runner.watchFactory = test.watcherFactory
-
-			err := runner.Dev(context.Background(), ioutil.Discard, nil)
-
-			testutil.CheckError(t, test.shouldErr, err)
-		})
-	}
-}
-
-func TestBuildAndDeployAllArtifacts(t *testing.T) {
-	builder := &TestBuilder{}
-	deployer := &TestDeployer{}
-	artifacts := []*latest.Artifact{
-		{ImageName: "image1"},
-		{ImageName: "image2"},
-	}
-
-	runner := createDefaultRunner(t)
-	runner.Builder = builder
-	runner.Deployer = deployer
-
-	ctx := context.Background()
-
-	// Both artifacts are changed
-	runner.watchFactory = NewWatcherFactory(nil, nil, []int{0, 1})
-	err := runner.Dev(ctx, ioutil.Discard, artifacts)
-
-	if err != nil {
-		t.Errorf("Didn't expect an error. Got %s", err)
-	}
-	if len(builder.built) != 2 {
-		t.Errorf("Expected 2 artifacts to be built. Got %d", len(builder.built))
-	}
-	if len(deployer.deployed) != 2 {
-		t.Errorf("Expected 2 artifacts to be deployed. Got %d", len(deployer.deployed))
-	}
-
-	// Only one is changed
-	runner.watchFactory = NewWatcherFactory(nil, nil, []int{1})
-	err = runner.Dev(ctx, ioutil.Discard, artifacts)
-
-	if err != nil {
-		t.Errorf("Didn't expect an error. Got %s", err)
-	}
-	if len(builder.built) != 1 {
-		t.Errorf("Expected 1 artifact to be built. Got %d", len(builder.built))
-	}
-	if len(deployer.deployed) != 2 {
-		t.Errorf("Expected 2 artifacts to be deployed. Got %d", len(deployer.deployed))
-	}
-}
-
-func TestShouldWatch(t *testing.T) {
-	var tests = []struct {
-		description   string
-		watch         []string
-		expectedMatch bool
-	}{
-		{
-			description:   "match all",
-			watch:         nil,
-			expectedMatch: true,
-		},
-		{
-			description:   "match full name",
-			watch:         []string{"domain/image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "match partial name",
-			watch:         []string{"image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "match any",
-			watch:         []string{"other", "image"},
-			expectedMatch: true,
-		},
-		{
-			description:   "no match",
-			watch:         []string{"other"},
-			expectedMatch: false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.description, func(t *testing.T) {
-			runner := &SkaffoldRunner{
-				opts: &config.SkaffoldOptions{
-					Watch: test.watch,
-				},
-			}
-
-			match := runner.shouldWatch(&latest.Artifact{
-				ImageName: "domain/image",
-			})
-
-			testutil.CheckDeepEqual(t, test.expectedMatch, match)
 		})
 	}
 }
